requirement.txt
astroid==2.2.5
Django==2.2
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
pkg-resources==0.0.0
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
pytz==2018.9
rope==0.14.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.1
wrapt==1.11.1
